requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
python-dotenv==0.20.0
Werkzeug==2.2.2
zipp==3.8.1
